requirements.txt
streamlit==1.30.0
tensorflow==2.13.0
opencv-python-headless==4.8.0.76
pandas==2.2.0
scikit-learn==1.3.0
matplotlib==3.8.0
kaggle==1.5.18
Pillow==10.0.0
tqdm==4.66.1
